Add files via upload
To logically analyze the user's feedback mechanism we propose to use Natural Language Processing (NLP) and Machine Learning(ML). NLP will identify the subjective
information of the feedback and gives the same as an input to ML. ML will further analyze this feedback data and try to explore the key focus area of the customer's problem and classifies it based on the magnitude of the problem and the emotions of the user.